[occm] remove instances v1 support
